- fixed bug in FileBlockStream.GetBlockBytes(int) : local variable was overflowing, - Console outputs were corrected and improved, - Output file is now deleted on exception while compressing\decompressing

+++ src/TestTask_GZipArchiver.Core/Models/FileBlockStream.cs
@@ -102,7 +102,7 @@ namespace TestTask_GZipArchiver.Core.Models
 
             byte[] result = new byte[blockSize];
 
-            var pos = BlockSize * blockNumber;
+            var pos = (long)BlockSize * blockNumber;
 
             lock (_locker)
             {
